requirements.txt
spotipy
python_dotenv
matplotlib